% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLindleyROC.R
\name{PLindleyROC}
\alias{PLindleyROC}
\alias{dPLD}
\title{Probability density function of Power Lindley Distribution}
\usage{
dPLD(x, alpha, beta)
}
\arguments{
\item{x}{vector of values}

\item{alpha}{shape parameter}

\item{beta}{scale parameter}
}
\value{
\code{dPLD()} returns a probability density function against the specified values.
}
\description{
\code{dPLD()} gives the probability density function of Power Lindley Distribution
}
\examples{
dPLD(c(1,2,3,4,5,200),3,2)
}
